Set HOME and userns in dev container
This commit adds a `containerEnv` setting that sets the `HOME`
variable. This is currently required due to `podman` setting
the home to `/root`, which causes permissions errors when
`vscode-server` tries to run.

Additionally, `--userns=keep-id` is added to the `runArgs`
setting. This is required to prevent permissions errors
when the `hass` command attempts to access files in
the `/workspace` mount.

ref:  https://github.com/microsoft/vscode-remote-release/issues/7657